Move graphql schema from data-access to data-access-plugin
