+++ packages/data-access-plugin/src/manager/schema/index.ts
@@ -1,9 +1,9 @@
 import GraphQLJSON from 'graphql-type-json';
 import { GraphQLDateTime } from 'graphql-iso-date';
 import * as a from 'apollo-server-express';
-import * as d from '@terascope/data-access';
 import * as query from './query';
 import * as mutation from './mutation';
+import { schema } from './misc';
 import { ManagerContext } from '../interfaces';
 
 const resolvers: a.IResolvers<any, ManagerContext> = {
@@ -17,7 +17,7 @@ export = a.makeExecutableSchema({
     typeDefs: [
         'scalar JSON',
         'scalar DateTime',
-        ...d.graphqlSchemas,
+        schema,
         query.schema,
         mutation.schema,
     ],
