--- packages/data-access/src/models/config/common.ts
@@ -1,4 +1,0 @@
-export const graphQLModel = `client_id: Int
-        id: ID!
-        created: DateTime!
-        updated: DateTime!`;
