+++ packages/data-access/src/models/roles.ts
@@ -1,13 +1,12 @@
 import * as es from 'elasticsearch';
 import { IndexModel, IndexModelOptions } from 'elasticsearch-store';
-import rolesConfig, { GraphQLSchema, Role } from './config/roles';
+import rolesConfig, { Role } from './config/roles';
 
 /**
  * Manager for Roles
 */
 export class Roles extends IndexModel<Role> {
     static IndexModelConfig = rolesConfig;
-    static GraphQLSchema = GraphQLSchema;
 
     constructor(client: es.Client, config: IndexModelOptions) {
         super(client, config, rolesConfig);
